MFA-Backend/mfa_app/requirements.txt
flask
flask_sqlalchemy
flask_migrate
flask-cors
bcrypt
pyjwt
python-dotenv
psycopg2-binary
opencv-contrib-python
numpy
flask_jwt_extended
